Fix model access in Julia after passing to thread
I don't understand the specifics, but I believe that passing the model into the ThreadState structure for each thread was moving where the BaseModel existed, or a pointer? The new API passes back a corrected pointer back to the Julia wrapper so get and set! still continue to work and not segfault